Add test for tunnelId extraction
Verify that extacting the tunnel id worked as expected.
